requirements.txt
anthropic>=0.39.0
feedparser>=6.0.11
requests>=2.31.0
beautifulsoup4>=4.12.0
PyYAML>=6.0.1
python-dotenv>=1.0.0
lxml>=5.0.0
